[TS] Pretty-print objects passed to console.log
