+++ crates/bindings-typescript/package.json
@@ -171,6 +171,7 @@
   "dependencies": {
     "base64-js": "^1.5.1",
     "headers-polyfill": "^4.0.3",
+    "object-inspect": "^1.13.4",
     "prettier": "^3.3.3",
     "pure-rand": "^7.0.1",
     "statuses": "^2.0.2",
@@ -179,9 +180,9 @@
   "peerDependencies": {
     "@tanstack/react-query": "^5.0.0",
     "react": "^18.0.0 || ^19.0.0-0 || ^19.0.0",
-    "vue": "^3.3.0",
     "svelte": "^4.0.0 || ^5.0.0",
-    "undici": "^6.19.2"
+    "undici": "^6.19.2",
+    "vue": "^3.3.0"
   },
   "peerDependenciesMeta": {
     "@tanstack/react-query": {
@@ -202,10 +203,10 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.17.0",
-    "svelte": "^5.0.0",
     "@size-limit/file": "^11.2.0",
     "@tanstack/react-query": "^5.90.19",
     "@types/fast-text-encoding": "^1.0.3",
+    "@types/object-inspect": "^1.13.0",
     "@types/react": "^19.1.13",
     "@types/statuses": "^2.0.6",
     "@typescript-eslint/eslint-plugin": "^8.18.2",
@@ -216,6 +217,7 @@
     "eslint-plugin-jsdoc": "^61.5.0",
     "globals": "^15.14.0",
     "size-limit": "^11.2.0",
+    "svelte": "^5.0.0",
     "ts-node": "^10.9.2",
     "tsup": "^8.1.0",
     "typescript": "^5.9.3",
