+++ crates/bindings-typescript/src/server/polyfills.ts
@@ -1,1 +1,4 @@
 import 'url-polyfill';
+import { console } from './console';
+
+globalThis.console = console;
