+++ crates/bindings-typescript/src/util-stub.ts
@@ -1,0 +1,1 @@
+export const inspect = {};
