+++ crates/bindings-typescript/tsup.config.ts
@@ -1,8 +1,8 @@
 // tsup.config.ts
 import { defineConfig, type Options } from 'tsup';
 
-function commonEsbuildTweaks() {
-  return (options: any) => {
+function commonEsbuildTweaks(): NonNullable<Options['esbuildOptions']> {
+  return options => {
     // Prefer "exports"."development" when deps provide it; harmless otherwise.
     options.conditions = ['development', 'import', 'default'];
     options.mainFields = ['browser', 'module', 'main'];
@@ -192,9 +192,14 @@ export default defineConfig([
       moduleSideEffects: ['src/server/polyfills.ts'],
     },
     external: ['undici', /^spacetime:sys.*$/],
-    noExternal: ['base64-js', 'statuses', 'pure-rand'],
+    noExternal: ['object-inspect', 'base64-js', 'statuses', 'pure-rand'],
     outExtension,
-    esbuildOptions: commonEsbuildTweaks(),
+    esbuildOptions: (options, ctx) => {
+      // object-inspect tries to import the node `util` module,
+      // which we want to disable.
+      options.alias = { util: './src/util-stub.ts' };
+      commonEsbuildTweaks()(options, ctx);
+    },
   },
 
   // TanStack subpath (SSR-friendly): dist/tanstack/index.{mjs,cjs}
